Shift data files around to clean things up
