+++ ios/evy/Utils/EVYFunctions.swift
@@ -9,7 +9,7 @@ import Foundation
 
 func evyCount(_ args: String) -> String {
     do {
-        let res = try EVYDataManager.i.parseProps(args)
+        let res = try EVYTextView.parseProps(args)
         switch res {
         case .array(let arrayValue):
             return String(arrayValue.count)
@@ -23,7 +23,7 @@ func evyCount(_ args: String) -> String {
 
 func evyFormatCurrency(_ args: String) -> String {
     do {
-        let res = try EVYDataManager.i.parseProps(args)
+        let res = try EVYTextView.parseProps(args)
         switch res {
         case .dictionary(let dictValue):
             guard let value = dictValue["value"] else {
